+++ QUBEKit/proteins/protein_run.py
@@ -3,7 +3,7 @@
 from QUBEKit.lennard_jones import LennardJones
 from QUBEKit.ligand import Protein
 from QUBEKit.parametrisation import XMLProtein
-from QUBEKit.proteins.protein_tools import get_water, pdb_reformat, qube_general
+from QUBEKit.proteins.protein_tools import get_water, pdb_reformat
 from QUBEKit.utils.file_handling import ExtractChargeData
 
 import argparse
