+++ QUBEKit/utils/decorators.py
@@ -6,6 +6,7 @@ from QUBEKit.utils.helpers import unpickle, COLOURS
 from functools import wraps
 import logging
 import os
+from typing import Callable, Optional
 
 
 def logger_format(log_file):
@@ -77,3 +78,32 @@ def exception_logger(func):
                     raise
 
     return wrapper
+
+
+def requires_package(package_name: str, conda_channel: Optional[str] = "conda-forge"):
+    """
+    This wraps a function to check if the required dependency is available before running.
+    """
+    from QUBEKit.utils.exceptions import DependencyError
+
+    def inner_decorator(function: Callable):
+        @wraps(function)
+        def wrapper(*args, **kwargs):
+            import importlib
+
+            try:
+                importlib.import_module(package_name)
+            except (ImportError, ModuleNotFoundError):
+                if conda_channel is None:
+                    install_method = f"pip install {package_name}."
+                else:
+                    install_method = f"conda install {package_name} -c {conda_channel}."
+                raise DependencyError(f"Missing dependecy for {package_name}. Try installing it with {install_method}")
+            except Exception as e:
+                raise e
+
+            return function(*args, **kwargs)
+
+        return wrapper
+
+    return inner_decorator
